% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/height.R
\name{calculate_h_dom}
\alias{calculate_h_dom}
\title{Calculate dominant height (h_dom) from top height and stand density}
\usage{
calculate_h_dom(h_top, N_at, c24, c53)
}
\arguments{
\item{h_top}{Top height in meters. Numeric scalar.}

\item{N_at}{Stand density (trees per hectare). Numeric scalar.}

\item{c24}{Model coefficient c24. Numeric scalar.}

\item{c53}{Model exponent c53. Numeric scalar.}
}
\value{
Dominant height \code{h_dom} in meters. Numeric scalar.
}
\description{
Computes the dominant height \code{h_dom} based on top height (\code{h_top}),
stand density (\code{N_at}), and model parameters \code{c24} and \code{c53}. The function
uses a piecewise formula:
}
\details{
\itemize{
\item If N_at > 250:
\deqn{h_dom = h_top - c24 * h_top^c53}
\item If 100 < N_at ≤ 250:
\deqn{h_dom = ((N_at - 100)/150) * (h_top - c24 * h_top^c53) + ((250 - N_at)/150) * h_top}
\item If N_at ≤ 100:
\deqn{h_dom = h_top}
}
}
\examples{
calculate_h_dom(h_top = 30, N_at = 300, c24 = 0.5, c53 = 1.2)
calculate_h_dom(h_top = 30, N_at = 180, c24 = 0.5, c53 = 1.2)
calculate_h_dom(h_top = 30, N_at = 80, c24 = 0.5, c53 = 1.2)

}
