% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/height.R
\name{calculate_hg_bt}
\alias{calculate_hg_bt}
\title{Calculate hg_bt (h) from h_at and thinning status}
\usage{
calculate_hg_bt(h_at, x_HD, c36L = 0.9885, c36H = 0.9746)
}
\arguments{
\item{h_at}{Adjusted height (numeric scalar). Typically from \code{calculate_adjusted_height()}.}

\item{x_HD}{Integer (0 or 1). Thinning indicator: 1 = thinning from above, 0 = otherwise.}

\item{c36L}{Coefficient for no thinning or thinning from below.}

\item{c36H}{Coefficient for thinning from above.}
}
\value{
Breast-height-adjusted height \code{h_bt} (numeric scalar).
}
\description{
This function computes hg_bt using hg_at and a thinning correction factor
that linearly blends between two coefficients (\code{c36L} and \code{c36H}) depending on the thinning type.
}
\details{
Formula:
\deqn{h_bt = ((c36L * (1 - x_HD)) + (c36H * x_HD)) * h_at}
}
\examples{
calculate_h_bt(h_at = 28.5, x_HD = 0, c36L = 0.9885, c36H = 0.9746)
calculate_h_bt(h_at = 28.5, x_HD = 1, c36L = 0.9885, c36H = 0.9746)

}
