% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basal_area.R
\name{calculate_ic_g_t7}
\alias{calculate_ic_g_t7}
\title{Calculate basal area increment for the year after passing t7}
\usage{
calculate_ic_g_t7(t, h_top_t1, h_top_t2, tgr_0, c)
}
\arguments{
\item{N}{Number of trees per hectare (numeric scalar or vector)}

\item{d}{Diameter at breast height (cm) (numeric scalar or vector)}
}
\value{
Basal area in square meters per hectare (numeric)
}
\description{
Formula from FORTRAN programma
}
\examples{
calculate_basal_area(N = 500, d = 30)
calculate_basal_area(N = c(500, 600), d = c(25, 30))

}
