% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stem_number.R
\name{calculate_N_at}
\alias{calculate_N_at}
\title{Calculate number of trees per hectare (N_at) from relative spacing and dominant height}
\usage{
calculate_N_at(S_percent, h_top)
}
\arguments{
\item{S_percent}{Relative spacing (\%) as a decimal (e.g. 0.25 for 25\%)}

\item{h_dom}{Dominant height in meters (numeric)}
}
\value{
Number of trees per hectare (N_at), numeric
}
\description{
This function computes the stand density (N_at) using the relative spacing percentage (S_percent)
and dominant height (h_dom), based on the spacing-density rule:
}
\details{
\deqn{N_at = (10746 / (S_percent * h_dom))^2}
}
\examples{
calculate_N_at(S_percent = 0.25, h_dom = 28)
calculate_N_at(S_percent = 0.30, h_dom = 22)

}
