% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/height.R
\name{calculate_t7}
\alias{calculate_t7}
\title{Calculate t7}
\usage{
calculate_t7(S_i, c1, c2, c3)
}
\arguments{
\item{S_i}{Site index (numeric scalar).}

\item{c1, }{c2, c3, Model coefficients (numeric scalars).}
}
\value{
Top height h_top at age t (numeric scalar).
}
\description{
Calculates the age at which the height is 7m
Inverse of formula 28, printed on p33
}
